refactor: Cambio de nombre al fragment de Jugadores
