+++ app/src/main/java/com/example/apptorneosajedrez/ui/jugadores/JugadoresViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.apptorneosajedrez.ui.slideshow
+package com.example.apptorneosajedrez.ui.jugadores
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class JugadoresViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is slideshow Fragment"
