+++ app/src/main/res/layout/fragment_jugadores.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.slideshow.JugadoresFragment">
+    tools:context=".ui.jugadores.JugadoresFragment">
 
     <ListView
         android:id="@+id/listViewJugadores"
