+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -19,7 +19,7 @@
 
     <fragment
         android:id="@+id/nav_jugadores"
-        android:name="com.example.apptorneosajedrez.ui.slideshow.JugadoresFragment"
+        android:name="com.example.apptorneosajedrez.ui.jugadores.JugadoresFragment"
         android:label="@string/menu_jugadores"
         tools:layout="@layout/fragment_jugadores" />
